The console runner's Listener implementations do not cross AppDomain boundaries.
