+++ src/Fixie.Console/Fixie.Console.x86.csproj
@@ -42,6 +42,7 @@
     <Compile Include="CommandLineParser.cs" />
     <Compile Include="ConsoleListener.cs" />
     <Compile Include="Foreground.cs" />
+    <Compile Include="ListenerFactory.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TeamCityListener.cs" />
