+++ src/Fixie.Tests/ConsoleRunner/TeamCityListenerTests.cs
@@ -13,8 +13,8 @@ namespace Fixie.Tests.ConsoleRunner
         public void ShouldReportResultsToTheConsoleInTeamCityFormat()
         {
             using (var console = new RedirectedConsole())
-            using (var listener = new TeamCityListener())
             {
+                var listener = new TeamCityListener();
                 var convention = SelfTestConvention.Build();
                 convention.CaseExecution.Skip(x => x.Method.Has<SkipAttribute>(), x => x.Method.GetCustomAttribute<SkipAttribute>().Reason);
